# spring-starter-api/src/main/resources/openApi.yaml
openapi: 3.0.3
info:
  title: Project Management API
  description: API for User, Project, and Task management with JWT and OAuth support.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server

# 1. GLOBAL SECURITY
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account and returns a JWT token. Public endpoint.
      operationId: registerUser # <--- METHOD NAME: registerUser(...)
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      summary: Login with credentials
      operationId: loginUser
      tags: [ Authentication ]
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie: # <--- This represents the HttpServletResponse action
              description: Contains the refresh token in a secure, HttpOnly cookie.
              schema:
                type: string
                example: refresh_token=eyJhbGci...; Path=/api/v1/auth/refresh; HttpOnly; Secure; SameSite=Strict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: refreshToken
      tags: [ Authentication ]
      parameters:
        - name: refresh_token
          in: cookie
          required: true
          schema:
            type: string
          description: The secure refresh token cookie
      responses:
        '200':
          description: New access token generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/login/oauth:
    post:
      summary: Login via OAuth Provider
      description: |
        Exchange an OAuth provider's ID token (e.g., from Google/GitHub) for an internal system JWT.
      operationId: loginViaOAuth # <--- METHOD NAME: loginViaOAuth(...)
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthLoginRequest'
      responses:
        '200':
          description: OAuth Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      summary: Logout user
      description: Logs out the current user by clearing the refresh token cookie.
      operationId: logoutUser
      tags: [ Authentication ]
      responses:
        '204':
          description: Logout successful, cookie cleared.
          headers:
            Set-Cookie:
              description: Clears the refresh_token cookie by setting its expiration to zero.
              schema:
                type: string
                example: refresh_token=; Path=/api/v1/auth/refresh; Max-Age=0; HttpOnly; Secure; SameSite=Strict
        '401':
          $ref: '#/components/responses/Unauthorized'
  # ----------------------------
  # USERS (PROTECTED)
  # ----------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication failed or token missing/invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string

    # --- Authentication ---
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    OAuthLoginRequest:
      type: object
      required: [provider, token]
      properties:
        provider:
          type: string
          enum: [google, github]
          description: The OAuth provider being used.
        token:
          type: string
          description: The ID token or Access token received from the provider (Frontend to Backend exchange).
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT Access Token
        refreshToken:
          type: string
          description: Refresh token
        tokenType:
          type: string
          example: "Bearer"
        expiresIn:
          type: integer
          description: Expiration time in seconds
        user:
          $ref: '#/components/schemas/UserResponse'

    # --- Roles ---
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string

    # --- Users ---
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        createdAt:
          type: string
          format: date-time

    # --- Projects ---
...